updates database var names as they didnt work for some reason
